backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic-settings==2.2.1
pydantic[email]==2.7.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
httpx==0.27.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
pyzbar==0.1.9
qrcode[pil]==7.4.2
reportlab==4.1.0
Pillow==10.3.0
openpyxl==3.1.2
aiofiles==23.2.1
